requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
langchain-chroma>=0.1.0
chromadb>=0.5.0

sentence-transformers>=2.2.2
transformers>=4.41.0
torch>=2.2.0

numpy>=1.24.0
scikit-learn>=1.3.0
tiktoken>=0.5.0

pypdf>=4.0.0

gradio>=4.36.0

python-dotenv>=1.0.0

pytest>=8.2.0
